creates intakeform folder with 3 seperate components for legibility
